refactor(nurse-doctor): extract workarea validation into WorkareaValidationService
Stage 1 of decomposing the 3,937-line workarea.component.ts god component.

Move the 8 required-field validators (checkMandatory, checkNurseRequirements,
checkCancerRequiredData, checkTMVisitDetailsRequiredData,
checkNCDScreeningRequiredData, checkQuickConsultDoctorData,
checkNCDScreeningHistory, checkPastObstericHistory — ~1,240 lines) into a new
stateless WorkareaValidationService. Each method takes (form, host); host is a
typed WorkareaValidationHost the component satisfies, so the same state and
service instances are used — behaviour is identical. The component keeps thin
delegators, so every call site is unchanged.

workarea.component.ts: 3937 -> 2730 lines. Bodies moved verbatim (verified: each
reverse-transforms byte-for-byte to the original). AOT build green.
