+++ src/app/app-modules/nurse-doctor/workarea/workarea.component.ts
@@ -56,6 +56,7 @@ import { SetLanguageComponent } from '../../core/components/set-language.compone
 import { Observable, Subscription, of } from 'rxjs';
 import { HttpServiceService } from '../../core/services/http-service.service';
 import { IdrsscoreService } from '../shared/services/idrsscore.service';
+import { WorkareaValidationService } from './workarea-validation.service';
 import { ZardDialogService } from 'Common-UI/v2/ui/dialog';
 import { environment } from 'src/environments/environment';
 import { CanComponentDeactivate } from '../../core/services/can-deactivate-guard.service';
@@ -275,8 +276,8 @@ export class WorkareaComponent
     private httpServiceService: HttpServiceService,
     private changeDetectorRef: ChangeDetectorRef,
     private masterdataService: MasterdataService,
-    private nurseService: NurseService,
-    private confirmationService: ConfirmationService,
+    public nurseService: NurseService,
+    public confirmationService: ConfirmationService,
     private doctorService: DoctorService,
     private route: ActivatedRoute,
     private beneficiaryDetailsService: BeneficiaryDetailsService,
@@ -284,7 +285,8 @@ export class WorkareaComponent
     private readonly viewContainerRef: ViewContainerRef,
     readonly sessionstorage: SessionStorageService,
     private idrsScoreService: IdrsscoreService,
-    private languageComponent: SetLanguageComponent
+    private languageComponent: SetLanguageComponent,
+    private readonly workareaValidation: WorkareaValidationService
   ) {}
   isSpecialist = false;
   doctorUpdateAndTCSubmit: any;
@@ -395,20 +397,7 @@ export class WorkareaComponent
   }
 
   checkMandatory() {
-    if (this.visitCategory === null || this.visitCategory === undefined) {
-      this.confirmationService.alert(
-        this.currentLanguageSet.alerts.info.proceedFurther
-      );
-    }
-    if (
-      this.nurseService.fileData !== undefined &&
-      this.nurseService.fileData.length > 0
-    ) {
-      this.confirmationService.alert(
-        this.currentLanguageSet.common.kindlyuploadthefiles
-      );
-      this.nurseService.fileData = null;
-    }
+    return this.workareaValidation.checkMandatory(this);
   }
 
   getVisitType() {
@@ -1448,621 +1437,11 @@ export class WorkareaComponent
   }
 
   checkNurseRequirements(medicalForm: any) {
-    const vitalsForm = <FormGroup>medicalForm.controls['patientVitalsForm'];
-    const examinationForm = <FormGroup>(
-      this.patientMedicalForm.controls['patientExaminationForm']
-    );
-    const pncForm = <FormGroup>(
-      this.patientMedicalForm.controls['patientPNCForm']
-    );
-    const ancForm = <FormGroup>(
-      this.patientMedicalForm.controls['patientANCForm']
-    );
-    const covidForm = <FormGroup>medicalForm.controls['patientVisitForm'];
-    const covidForm2 = <FormGroup>covidForm.controls['patientCovidForm'];
-    const referForm = <FormGroup>medicalForm.controls['patientReferForm'];
-    const historyForm = <FormGroup>medicalForm.controls['patientHistoryForm'];
-
-    const required = [];
-
-    if (environment.isMMUOfflineSync) {
-      if (
-        this.enableLungAssessment === true &&
-        this.beneficiaryAge >= 18 &&
-        this.nurseService.isAssessmentDone === false
-      ) {
-        required.push('Please perform Lung Assessment');
-      }
-    }
-
-    console.log('pncForm', pncForm);
-
-    if (this.visitCategory === 'PNC') {
-      if (pncForm.controls['deliveryPlace'].errors) {
-        required.push(this.currentLanguageSet.pncData.placeofDelivery);
-      }
-      if (pncForm.controls['deliveryType'].errors) {
-        required.push(this.currentLanguageSet.pncData.typeofDelivery);
-      }
-    }
-
-    if (this.visitCategory === 'ANC') {
-      const ancdetailsForm = <FormGroup>(
-        ancForm.controls['patientANCDetailsForm']
-      );
-      const ANCVitalsForm = <FormGroup>(
-        medicalForm.controls['patientVitalsForm']
-      );
-      console.log('ANCCaseRecordForm', ANCVitalsForm);
-      if (ancdetailsForm.controls['primiGravida'].errors) {
-        required.push(
-          this.currentLanguageSet.ancData.ancDataDetails.primiGravida
-        );
-      }
-      if (ancdetailsForm.controls['lmpDate'].errors) {
-        required.push(
-          this.currentLanguageSet.ancData.ancDataDetails.lastMenstrualPeriod
-        );
-      }
-
-      if (this.attendantType === 'doctor') {
-        const ANCCaseRecordForm = <FormGroup>(
-          medicalForm.controls['patientCaseRecordForm']
-        );
-        if (this.rbsPresent > 0) {
-          let investigationCount = 0;
-          const labTestArray =
-            ANCCaseRecordForm.controls['generalDoctorInvestigationForm'].value
-              .labTest;
-          if (
-            labTestArray !== undefined &&
-            labTestArray !== null &&
-            labTestArray.length > 0
-          ) {
-            labTestArray.forEach((element: any) => {
-              if (
-                element.procedureName !== null &&
-                element.procedureName.toLowerCase() ===
-                  environment.RBSTest.toLowerCase()
-              ) {
-                investigationCount++;
-              }
-            });
-          }
-
-          if (
-            investigationCount === 0 &&
-            ANCVitalsForm.controls['rbsTestResult'].value === null
-          ) {
-            required.push(
-              this.currentLanguageSet.pleaseSelectRBSTestInInvestigation
-            );
-          }
-        }
-        if (this.heamoglobinPresent > 0) {
-          let investigationCount = 0;
-          const labTestArray =
-            ANCCaseRecordForm.controls['generalDoctorInvestigationForm'].value
-              .labTest;
-          if (
-            labTestArray !== null &&
-            labTestArray !== undefined &&
-            labTestArray.length > 0
-          ) {
-            labTestArray.forEach((element: any) => {
-              if (
-                element.procedureName !== null &&
-                element.procedureName.toLowerCase() ===
-                  environment.haemoglobinTest.toLowerCase()
-              ) {
-                investigationCount++;
-              }
-            });
-          }
-
-          if (investigationCount === 0) {
-            required.push(
-              this.currentLanguageSet.pleaseSelectHeamoglobinTestInInvestigation
-            );
-          }
-        }
-      }
-    }
-
-    if (this.visitCategory !== 'General OPD (QC)') {
-      const pregForm = <FormGroup>medicalForm.controls['patientHistoryForm'];
-      const pregForm1 = <FormGroup>pregForm.controls['pastObstericHistory'];
-      const pregForm2 = <FormGroup>(
-        pregForm1.controls['pastObstericHistoryList']
-      );
-      if (this.attendantType === 'nurse') {
-        if (pregForm2.controls) {
-          const score1 = Number(pregForm2.controls['length']);
-          for (let i = 0; i < score1; i++) {
-            const pregForm3 = <FormGroup>pregForm2.controls[i];
-            if (
-              pregForm3.controls['pregOutcome'].value &&
-              pregForm3.controls['pregOutcome'].value.pregOutcome === 'Abortion'
-            ) {
-              if (
-                pregForm3.controls['abortionType'].value &&
-                pregForm3.controls['abortionType'].value.complicationValue ===
-                  'Induced' &&
-                pregForm3.controls['typeofFacility'].errors
-              ) {
-                required.push(
-                  this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                    .typeofFacility +
-                    '-' +
-                    this.currentLanguageSet.historyData.opdNCDPNCHistory
-                      .obstetric.orderofPregnancy +
-                    ' ' +
-                    pregForm3.value.pregOrder
-                );
-              }
-              if (pregForm3.controls['postAbortionComplication'].errors) {
-                required.push(
-                  this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                    .complicationPostAbortion +
-                    '-' +
-                    this.currentLanguageSet.historyData.opdNCDPNCHistory
-                      .obstetric.orderofPregnancy +
-                    ' ' +
-                    pregForm3.value.pregOrder
-                );
-              }
-              if (pregForm3.controls['abortionType'].errors) {
-                required.push(
-                  this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                    .typeOfAbortion +
-                    '-' +
-                    this.currentLanguageSet.historyData.opdNCDPNCHistory
-                      .obstetric.orderofPregnancy +
-                    ' ' +
-                    pregForm3.value.pregOrder
-                );
-              }
-              if (pregForm3.controls['pregDuration'].errors) {
-                required.push(
-                  this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                    .noOfcompletedWeeks +
-                    '-' +
-                    this.currentLanguageSet.historyData.opdNCDPNCHistory
-                      .obstetric.orderofPregnancy +
-                    ' ' +
-                    pregForm3.value.pregOrder
-                );
-              }
-            }
-          }
-        }
-      }
-    }
-
-    if (this.visitCategory === 'COVID-19 Screening') {
-      const historyForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientHistoryForm']
-      );
-      console.log('HistoryForm', historyForm);
-      const historyForm2 = <FormGroup>(
-        historyForm.controls['comorbidityHistory']
-      );
-      const historyForm3 = <FormArray>(
-        historyForm2.controls['comorbidityConcurrentConditionsList']
-      );
-      const historyForm4 = <FormGroup>historyForm3.controls[0];
-      if (historyForm4.controls['comorbidConditions'].errors) {
-        required.push(
-          this.currentLanguageSet.historyData.ancHistory
-            .combordityANC_OPD_NCD_PNC.comorbidConditions
-        );
-      }
-      if (covidForm2.controls['contactStatus'].errors) {
-        required.push(this.currentLanguageSet.contactHistory);
-      }
-
-      if (covidForm2.controls['travelStatus'].errors) {
-        required.push(this.currentLanguageSet.covid.travelHistory);
-      }
-      if (covidForm2.controls['symptom'].errors) {
-        required.push(
-          this.currentLanguageSet.ExaminationData.cancerScreeningExamination
-            .symptoms.symptoms
-        );
-      }
-    }
-    if (
-      this.visitCategory === 'General OPD' &&
-      this.attendantType === 'doctor'
-    ) {
-      const diagForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientCaseRecordForm']
-      );
-      const diagForm1 = <FormGroup>diagForm.controls['generalDiagnosisForm'];
-      const diagForm2 = <FormArray>(
-        diagForm1.controls['provisionalDiagnosisList']
-      );
-      const diagForm3 = <FormGroup>diagForm2.controls[0];
-      if (diagForm3.controls['provisionalDiagnosis'].errors) {
-        required.push(
-          this.currentLanguageSet.DiagnosisDetails.provisionaldiagnosis
-        );
-      }
-
-      if (!diagForm3.controls['provisionalDiagnosis'].errors) {
-        diagForm2.value.filter((item: any) => {
-          if (
-            item.provisionalDiagnosis &&
-            (item.conceptID === null ||
-              item.conceptID === undefined ||
-              item.conceptID === '')
-          )
-            required.push(
-              this.currentLanguageSet.provisionalDiagnosisIsNotValid
-            );
-        });
-      }
-    }
-    if (this.visitCategory === 'PNC' && this.attendantType === 'doctor') {
-      const diagForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientCaseRecordForm']
-      );
-      const diagForm1 = <FormGroup>diagForm.controls['generalDiagnosisForm'];
-      const diagForm2 = <FormArray>(
-        diagForm1.controls['provisionalDiagnosisList']
-      );
-      const diagForm3 = <FormGroup>diagForm2.controls[0];
-      if (diagForm3.controls['provisionalDiagnosis'].errors) {
-        required.push(
-          this.currentLanguageSet.DiagnosisDetails.provisionaldiagnosis
-        );
-      }
-
-      if (!diagForm3.controls['provisionalDiagnosis'].errors) {
-        diagForm2.value.filter((item: any) => {
-          if (
-            item.provisionalDiagnosis &&
-            (item.conceptID === null ||
-              item.conceptID === undefined ||
-              item.conceptID === '')
-          )
-            required.push(
-              this.currentLanguageSet.provisionalDiagnosisIsNotValid
-            );
-        });
-      }
-
-      const confirmatorydiagForm = <FormArray>(
-        diagForm1.controls['confirmatoryDiagnosisList']
-      );
-
-      confirmatorydiagForm.value.filter((item: any) => {
-        if (
-          item.confirmatoryDiagnosis &&
-          (item.conceptID === null ||
-            item.conceptID === undefined ||
-            item.conceptID === '')
-        )
-          required.push(
-            this.currentLanguageSet.confirmatoryDiagnosisIsNotValid
-          );
-      });
-    }
-    if (
-      this.visitCategory === 'Cancer Screening' &&
-      this.attendantType === 'doctor'
-    ) {
-      const diagForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientCaseRecordForm']
-      );
-      const diagForm1 = <FormGroup>diagForm.controls['diagnosisForm'];
-      if (diagForm1.controls['provisionalDiagnosisPrimaryDoctor'].errors) {
-        required.push(
-          this.currentLanguageSet.DiagnosisDetails.provisionaldiagnosis
-        );
-      }
-    }
-    if (
-      this.visitCategory === 'COVID-19 Screening' &&
-      this.attendantType === 'doctor'
-    ) {
-      const diagForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientCaseRecordForm']
-      );
-      const diagForm1 = <FormGroup>diagForm.controls['generalDiagnosisForm'];
-      console.log('diag', diagForm1);
-      if (diagForm1.controls['doctorDiagnosis'].errors) {
-        required.push(this.currentLanguageSet.doctorDiagnosis);
-      }
-    }
-
-    if (this.visitCategory !== 'General OPD (QC)') {
-      const personalHistory = historyForm.controls['personalHistory'];
-      const allergyList = personalHistory.value.allergicList;
-
-      let snomedTermNotMapped = false;
-
-      if (allergyList.length > 0) {
-        for (let i = 0; i < allergyList.length; i++) {
-          if (allergyList[i].allergyType !== null) {
-            if (
-              allergyList[i].snomedCode === null &&
-              allergyList[i].snomedTerm !== null
-            ) {
-              snomedTermNotMapped = true;
-            } else if (
-              allergyList[i].snomedCode !== null &&
-              allergyList[i].snomedTerm === null
-            ) {
-              snomedTermNotMapped = true;
-            }
-          }
-        }
-      }
-
-      if (snomedTermNotMapped) {
-        required.push(this.currentLanguageSet.allergyNameIsNotValid);
-      }
-    }
-    if (vitalsForm !== undefined && vitalsForm !== null) {
-      if (vitalsForm.controls['systolicBP_1stReading'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-            .systolicBP
-        );
-      }
-      if (vitalsForm.controls['diastolicBP_1stReading'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-            .diastolicBP
-        );
-      }
-
-      if (vitalsForm.controls['height_cm'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.AnthropometryDataANC_OPD_NCD_PNC
-            .height
-        );
-      }
-      if (vitalsForm.controls['weight_Kg'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.AnthropometryDataANC_OPD_NCD_PNC
-            .weight
-        );
-      }
-      if (vitalsForm.controls['temperature'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-            .temperature
-        );
-      }
-      if (vitalsForm.controls['pulseRate'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-            .pulseRate
-        );
-      }
-    }
-    if (this.visitCategory === 'NCD care') {
-      const diagnosisForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientCaseRecordForm']
-      );
-      if (diagnosisForm !== undefined) {
-        const diagnosisForm1 = <FormGroup>(
-          diagnosisForm.controls['generalDiagnosisForm']
-        );
-        if (diagnosisForm1 !== undefined) {
-          const temp =
-            diagnosisForm1.controls['ncdScreeningConditionArray'].value;
-          if (diagnosisForm1.controls['ncdScreeningConditionArray'].errors) {
-            required.push(this.currentLanguageSet.casesheet.ncdCondition);
-          }
-          let flag = false;
-          if (temp !== undefined && temp !== null && temp.length > 0) {
-            temp.forEach((element: any) => {
-              if (element === 'Other') flag = true;
-            });
-          }
-
-          console.log(
-            diagnosisForm1.controls['ncdScreeningConditionOther'].value
-          );
-          if (
-            flag &&
-            diagnosisForm1.controls['ncdScreeningConditionOther'].value === null
-          )
-            required.push(this.currentLanguageSet.nCDConditionOther);
-        }
-      }
-    }
-    console.log('referForm', referForm);
-    if (this.attendantType === 'doctor') {
-      const referForm = <FormGroup>medicalForm.controls['patientReferForm'];
-      if (
-        referForm.controls['referredToInstituteName'].value === null &&
-        sessionStorage.getItem('instFlag') === 'true' &&
-        sessionStorage.getItem('suspectFlag') === 'true'
-      ) {
-        required.push(
-          'this.currentLanguageSet.Referdetails.higherhealthcarecenter'
-        );
-      }
-
-      if (referForm.controls['refrredToAdditionalServiceList'].value !== null) {
-        if (
-          referForm.controls['refrredToAdditionalServiceList'].value.length > 0
-        ) {
-          if (referForm.controls['referralReason'].errors) {
-            required.push(this.currentLanguageSet.Referdetails.referralReason);
-          }
-        } else if (
-          referForm.controls['referredToInstituteName'].value !== null
-        ) {
-          if (referForm.controls['referralReason'].errors) {
-            required.push(this.currentLanguageSet.Referdetails.referralReason);
-          }
-        }
-      } else if (referForm.controls['referredToInstituteName'].value !== null) {
-        if (referForm.controls['referralReason'].errors) {
-          required.push(this.currentLanguageSet.Referdetails.referralReason);
-        }
-      }
-    }
-    console.log(examinationForm, 'examinationForm');
-    if (examinationForm !== undefined && examinationForm !== null) {
-      const generalExaminationForm = <FormGroup>(
-        examinationForm.controls['generalExaminationForm']
-      );
-      if (generalExaminationForm.controls['typeOfDangerSigns'].errors) {
-        required.push(
-          this.currentLanguageSet.ExaminationData.ANC_OPD_PNCExamination
-            .genExamination.dangersigns
-        );
-      }
-      if (generalExaminationForm.controls['lymphnodesInvolved'].errors) {
-        required.push(
-          this.currentLanguageSet.ExaminationData.ANC_OPD_PNCExamination
-            .genExamination.lymph
-        );
-      }
-      if (generalExaminationForm.controls['typeOfLymphadenopathy'].errors) {
-        required.push(
-          this.currentLanguageSet.ExaminationData.ANC_OPD_PNCExamination
-            .genExamination.typeofLymphadenopathy
-        );
-      }
-      if (generalExaminationForm.controls['extentOfEdema'].errors) {
-        required.push(
-          this.currentLanguageSet.ExaminationData.ANC_OPD_PNCExamination
-            .genExamination.extentofEdema
-        );
-      }
-      if (generalExaminationForm.controls['edemaType'].errors) {
-        required.push(
-          this.currentLanguageSet.ExaminationData.ANC_OPD_PNCExamination
-            .genExamination.typeofEdema
-        );
-      }
-    }
-
-    // Ensure doctor has added at least one prescription
-    if (this.attendantType === 'doctor') {
-      try {
-        const caseRecordForm = <FormGroup>(
-          medicalForm.controls['patientCaseRecordForm']
-        );
-        const drugPrescriptionForm = <FormGroup>(
-          (caseRecordForm && caseRecordForm.controls
-            ? caseRecordForm.controls['drugPrescriptionForm']
-            : null)
-        );
-        if (drugPrescriptionForm) {
-          let prescribedDrugs =
-            drugPrescriptionForm.value &&
-            drugPrescriptionForm.value.prescribedDrugs
-              ? drugPrescriptionForm.value.prescribedDrugs
-              : [];
-          prescribedDrugs = prescribedDrugs.filter((d: any) => !!d.createdBy);
-          if (!prescribedDrugs || prescribedDrugs.length === 0) {
-            required.push(
-              this.currentLanguageSet?.Prescription?.prescriptionRequired ||
-                'Please add at least one prescription'
-            );
-          }
-        }
-      } catch (err) {
-        console.warn('Error validating prescription presence', err);
-      }
-    }
-
-    if (required.length) {
-      this.confirmationService.notify(
-        this.currentLanguageSet.alerts.info.mandatoryFields,
-        required
-      );
-      this.resetSpinnerandEnableTheSubmitButton();
-      return 0;
-    } else {
-      return 1;
-    }
+    return this.workareaValidation.checkNurseRequirements(medicalForm, this);
   }
 
   checkCancerRequiredData(medicalForm: any) {
-    const vitalsForm = <FormGroup>medicalForm.controls['patientVitalsForm'];
-    const referForm = <FormGroup>medicalForm.controls['patientReferForm'];
-    const required = [];
-
-    if (vitalsForm !== undefined && vitalsForm !== null) {
-      if (vitalsForm.controls['height_cm'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.AnthropometryDataANC_OPD_NCD_PNC
-            .height
-        );
-      }
-      if (vitalsForm.controls['weight_Kg'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.AnthropometryDataANC_OPD_NCD_PNC
-            .weight
-        );
-      }
-      if (vitalsForm.controls['systolicBP_1stReading'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-            .systolicBP
-        );
-      }
-      if (vitalsForm.controls['diastolicBP_1stReading'].errors) {
-        required.push(
-          this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-            .diastolicBP
-        );
-      }
-    }
-
-    if (
-      this.visitCategory === 'Cancer Screening' &&
-      this.attendantType === 'doctor'
-    ) {
-      const diagForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientCaseRecordForm']
-      );
-      if (diagForm.controls['provisionalDiagnosisPrimaryDoctor'].errors) {
-        required.push(
-          this.currentLanguageSet.DiagnosisDetails.provisionaldiagnosis
-        );
-      }
-    }
-
-    if (this.attendantType === 'doctor') {
-      if (referForm.controls['refrredToAdditionalServiceList'].value !== null) {
-        if (
-          referForm.controls['refrredToAdditionalServiceList'].value.length > 0
-        ) {
-          if (referForm.controls['referralReason'].errors) {
-            required.push(this.currentLanguageSet.Referdetails.referralReason);
-          }
-        } else if (referForm.controls['referredToInstituteID'].value !== null) {
-          if (referForm.controls['referralReason'].errors) {
-            required.push(this.currentLanguageSet.Referdetails.referralReason);
-          }
-        }
-      } else if (referForm.controls['referredToInstituteID'].value !== null) {
-        if (referForm.controls['referralReason'].errors) {
-          required.push(this.currentLanguageSet.Referdetails.referralReason);
-        }
-      }
-    }
-
-    if (required.length) {
-      this.confirmationService.notify(
-        this.currentLanguageSet.alerts.info.mandatoryFields,
-        required
-      );
-      this.resetSpinnerandEnableTheSubmitButton();
-      return false;
-    } else {
-      return true;
-    }
+    return this.workareaValidation.checkCancerRequiredData(medicalForm, this);
   }
   submitTMPatientVisitForm(medicalForm: any) {
     if (this.checkTMVisitDetailsRequiredData(medicalForm)) {
@@ -2106,315 +1485,16 @@ export class WorkareaComponent
     }
   }
   checkTMVisitDetailsRequiredData(medicalForm: any) {
-    const required = [];
-    const tmVisitForm = <FormGroup>medicalForm.controls['patientVisitForm'];
-    const tmVisitForm2 = <FormGroup>tmVisitForm.controls['tmcConfirmationForm'];
-    if (tmVisitForm2.controls['refrredToAdditionalServiceList'].errors) {
-      required.push(this.currentLanguageSet.Referdetails.referredtoinstitute);
-    }
-    console.log('tmVisitForm2', tmVisitForm2);
-    if (
-      tmVisitForm2.value !== undefined &&
-      tmVisitForm2.value.isTMCConfirmed !== undefined &&
-      tmVisitForm2.value.isTMCConfirmed === true
-    )
-      tmVisitForm2.patchValue({ refrredToAdditionalServiceList: null });
-    if (tmVisitForm2.controls['tmcConfirmed'].errors) {
-      required.push(this.currentLanguageSet.tmcConfirmed);
-    }
-    console.log('tmVisitForm2', tmVisitForm2);
-    if (required.length) {
-      this.confirmationService.notify(
-        this.currentLanguageSet.alerts.info.mandatoryFields,
-        required
-      );
-      return false;
-    } else {
-      return true;
-    }
+    return this.workareaValidation.checkTMVisitDetailsRequiredData(
+      medicalForm,
+      this
+    );
   }
   checkNCDScreeningRequiredData(medicalForm: any) {
-    //WDF requirement
-    // let NCDScreeningForm = <FormGroup>medicalForm.controls['NCDScreeningForm'];
-    const NCDScreeningForm = <FormGroup>(
-      medicalForm.controls['patientVitalsForm']
+    return this.workareaValidation.checkNCDScreeningRequiredData(
+      medicalForm,
+      this
     );
-    const ncdIDRSScreeningForm = <FormGroup>(
-      medicalForm.controls['idrsScreeningForm']
-    );
-    const required = [];
-
-    if (environment.isMMUOfflineSync) {
-      if (
-        this.enableLungAssessment === true &&
-        this.beneficiaryAge >= 18 &&
-        this.nurseService.isAssessmentDone === false
-      ) {
-        required.push('Please perform Lung Assessment');
-      }
-    }
-
-    let count = 0;
-    const physicalActivityMandatory = <FormGroup>(
-      medicalForm.controls['patientHistoryForm'].controls[
-        'physicalActivityHistory'
-      ]
-    );
-    if (
-      this.attendantType === 'nurse' &&
-      this.diabetesSelected === 1 &&
-      NCDScreeningForm.controls['rbsCheckBox'].value === true &&
-      NCDScreeningForm.controls['rbsTestResult'].value === null
-    ) {
-      required.push('Please perform RBS Test under Vitals');
-    }
-    if (this.beneficiary.ageVal >= 30) {
-      const familyDiseaseList =
-        medicalForm.controls.patientHistoryForm.controls.familyHistory.controls
-          .familyDiseaseList.value;
-      familyDiseaseList.forEach((element: any) => {
-        if (
-          element.diseaseType !== null &&
-          element.deleted === false &&
-          element.diseaseType.diseaseType === 'Diabetes Mellitus'
-        ) {
-          count++;
-        }
-      });
-      if (count === 0) {
-        required.push(
-          this.currentLanguageSet.pleaseSelectDiabetesMellitusInFamilyHistory
-        );
-      }
-      if (physicalActivityMandatory.controls['activityType'].errors) {
-        required.push(this.currentLanguageSet.physicalActivity);
-      }
-    }
-    let familyMember = 0;
-    const familyDiseasesList =
-      medicalForm.controls.patientHistoryForm.controls.familyHistory.controls
-        .familyDiseaseList.value;
-    let familyDiseasesLength = familyDiseasesList.length;
-    for (let element = 0; element < familyDiseasesList.length; element++) {
-      if (
-        familyDiseasesList[element].diseaseType !== null &&
-        familyDiseasesList[element].deleted === false
-      ) {
-        if (
-          familyDiseasesList[element].familyMembers !== null &&
-          familyDiseasesList[element].familyMembers.length > 0
-        ) {
-          familyMember++;
-        }
-      } else {
-        familyDiseasesLength--;
-      }
-    }
-    if (familyMember !== familyDiseasesLength) {
-      required.push(this.currentLanguageSet.familyMemberInFamilyHistory);
-    }
-    if (ncdIDRSScreeningForm.controls['requiredList'].value !== null) {
-      const ar = ncdIDRSScreeningForm.controls['requiredList'].value;
-      for (let i = 0; i < ar.length; i++) {
-        if (ar[i] !== 'Hypertension') {
-          required.push(ar[i]);
-        }
-      }
-    }
-
-    //WDF requirement -> to check whether RBS test is prescribed or not
-    if (this.attendantType === 'doctor') {
-      const NCDScreeningCaseRecordForm = <FormGroup>(
-        medicalForm.controls['patientCaseRecordForm']
-      );
-      if (this.rbsPresent > 0 && this.diabetesSelected > 0) {
-        let investigationCount = 0;
-        const labTestArray =
-          NCDScreeningCaseRecordForm.controls['generalDoctorInvestigationForm']
-            .value.labTest;
-        if (
-          labTestArray !== undefined &&
-          labTestArray !== null &&
-          labTestArray.length > 0
-        ) {
-          labTestArray.forEach((element: any) => {
-            if (
-              element.procedureName !== null &&
-              element.procedureName.toLowerCase() ===
-                environment.RBSTest.toLowerCase()
-            ) {
-              investigationCount++;
-            }
-          });
-        }
-
-        if (
-          investigationCount === 0 &&
-          this.diabetesSelected === 1 &&
-          NCDScreeningForm.controls['rbsCheckBox'].value === true &&
-          NCDScreeningForm.controls['rbsTestResult'].value === null
-        ) {
-          required.push('Please select RBS Test under Vitals or Investigation');
-        }
-        if (
-          investigationCount === 0 &&
-          this.diabetesSelected === 1 &&
-          NCDScreeningForm.controls['rbsCheckBox'].value === false &&
-          NCDScreeningForm.controls['rbsTestResult'].value === null
-        ) {
-          required.push('Please select RBS Test under Investigation');
-        }
-      }
-      if (this.visualAcuityPresent > 0 && this.visualAcuityMandatory > 0) {
-        let investigationVisualCount = 0;
-        const labTestArray =
-          NCDScreeningCaseRecordForm.controls['generalDoctorInvestigationForm']
-            .value.labTest;
-        if (
-          labTestArray !== null &&
-          labTestArray !== undefined &&
-          labTestArray.length > 0
-        ) {
-          labTestArray.forEach((element: any) => {
-            if (
-              element.procedureName !== null &&
-              element.procedureName.toLowerCase() ===
-                environment.visualAcuityTest.toLowerCase()
-            ) {
-              investigationVisualCount++;
-            }
-          });
-        }
-
-        if (investigationVisualCount === 0) {
-          required.push(
-            this.currentLanguageSet.pleaseSelectVisualAcuityTestInInvestigation
-          );
-        }
-      }
-    }
-    //WDF requirement
-    if (NCDScreeningForm.controls['height_cm'].errors) {
-      required.push(
-        this.currentLanguageSet.vitalsDetails.AnthropometryDataANC_OPD_NCD_PNC
-          .height
-      );
-    }
-    if (NCDScreeningForm.controls['weight_Kg'].errors) {
-      required.push(
-        this.currentLanguageSet.vitalsDetails.AnthropometryDataANC_OPD_NCD_PNC
-          .weight
-      );
-    }
-    if (NCDScreeningForm.controls['waistCircumference_cm'].errors) {
-      required.push(
-        this.currentLanguageSet.vitalsDetails.vitalsCancerscreening_QC
-          .waistCircumference
-      );
-    }
-    console.log('ncdTemp', this.ncdTemperature);
-    if (
-      NCDScreeningForm.controls['temperature'].errors &&
-      this.ncdTemperature === true
-    ) {
-      required.push(
-        this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-          .temperature
-      );
-    }
-    if (NCDScreeningForm.controls['pulseRate'].errors) {
-      required.push(
-        this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-          .pulseRate
-      );
-    }
-    if (NCDScreeningForm.controls['systolicBP_1stReading'].errors) {
-      required.push(
-        this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-          .systolicBP
-      );
-    }
-    if (NCDScreeningForm.controls['diastolicBP_1stReading'].errors) {
-      required.push(
-        this.currentLanguageSet.vitalsDetails.vitalsDataANC_OPD_NCD_PNC
-          .diastolicBP
-      );
-    }
-    if (this.attendantType === 'doctor') {
-      const diagForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientCaseRecordForm']
-      );
-      const diagForm1 = <FormGroup>diagForm.controls['generalDiagnosisForm'];
-      const diagForm2 = <FormArray>(
-        diagForm1.controls['provisionalDiagnosisList']
-      );
-      const diagForm3 = <FormGroup>diagForm2.controls[0];
-      if (
-        diagForm3.controls['provisionalDiagnosis'].errors &&
-        this.enableProvisionalDiag === true
-      ) {
-        required.push(
-          this.currentLanguageSet.DiagnosisDetails.provisionaldiagnosis
-        );
-      }
-
-      if (!diagForm3.controls['provisionalDiagnosis'].errors) {
-        diagForm2.value.filter((item: any) => {
-          if (
-            item.provisionalDiagnosis &&
-            (item.conceptID === null ||
-              item.conceptID === undefined ||
-              item.conceptID === '')
-          )
-            required.push(
-              this.currentLanguageSet.provisionalDiagnosisIsNotValid
-            );
-        });
-      }
-    }
-
-    if (this.attendantType === 'doctor') {
-      const referForm = <FormGroup>medicalForm.controls['patientReferForm'];
-      if (
-        referForm.controls['referredToInstituteName'].value === null &&
-        sessionStorage.getItem('instFlag') === 'true' &&
-        sessionStorage.getItem('suspectFlag') === 'true'
-      ) {
-        required.push(
-          this.currentLanguageSet.Referdetails.higherhealthcarecenter
-        );
-      }
-      if (referForm.controls['refrredToAdditionalServiceList'].value !== null) {
-        if (
-          referForm.controls['refrredToAdditionalServiceList'].value.length > 0
-        ) {
-          if (referForm.controls['referralReason'].errors) {
-            required.push(this.currentLanguageSet.Referdetails.referralReason);
-          }
-        } else if (
-          referForm.controls['referredToInstituteName'].value !== null
-        ) {
-          if (referForm.controls['referralReason'].errors) {
-            required.push(this.currentLanguageSet.Referdetails.referralReason);
-          }
-        }
-      } else if (referForm.controls['referredToInstituteName'].value !== null) {
-        if (referForm.controls['referralReason'].errors) {
-          required.push(this.currentLanguageSet.Referdetails.referralReason);
-        }
-      }
-    }
-
-    if (required.length) {
-      this.confirmationService.notify(
-        this.currentLanguageSet.alerts.info.mandatoryFields,
-        required
-      );
-      this.resetSpinnerandEnableTheSubmitButton();
-      return false;
-    } else {
-      return true;
-    }
   }
 
   /**
@@ -2443,127 +1523,10 @@ export class WorkareaComponent
   }
 
   checkQuickConsultDoctorData(patientMedicalForm: any) {
-    const form = <FormGroup>(
-      this.patientMedicalForm.controls['patientQuickConsultForm']
+    return this.workareaValidation.checkQuickConsultDoctorData(
+      patientMedicalForm,
+      this
     );
-
-    const referForm = <FormGroup>(
-      patientMedicalForm.controls['patientReferForm']
-    );
-
-    const required = [];
-
-    if (form.controls['chiefComplaintList'].errors) {
-      required.push(
-        this.currentLanguageSet.nurseData.chiefComplaintsDetails.chiefComplaints
-      );
-    }
-    if (form.controls['clinicalObservation'].errors) {
-      required.push(this.currentLanguageSet.casesheet.clinicalObs);
-    }
-    if (form.controls['provisionalDiagnosisList'].errors) {
-      required.push(
-        this.currentLanguageSet.DiagnosisDetails.provisionaldiagnosis
-      );
-    }
-    if (
-      this.visitCategory === 'General OPD (QC)' &&
-      this.attendantType === 'doctor'
-    ) {
-      const diagForm = <FormGroup>(
-        this.patientMedicalForm.controls['patientQuickConsultForm']
-      );
-      const diagForm2 = <FormArray>(
-        diagForm.controls['provisionalDiagnosisList']
-      );
-      console.log('diagForm2', diagForm2);
-      const diagForm3 = <FormGroup>diagForm2.controls[0];
-      if (diagForm3.controls['provisionalDiagnosis'].errors) {
-        required.push(
-          this.currentLanguageSet.DiagnosisDetails.provisionaldiagnosis
-        );
-      }
-
-      if (!diagForm3.controls['provisionalDiagnosis'].errors) {
-        diagForm2.value.filter((item: any) => {
-          if (
-            item.provisionalDiagnosis &&
-            (item.conceptID === null ||
-              item.conceptID === undefined ||
-              item.conceptID === '')
-          )
-            required.push(
-              this.currentLanguageSet.provisionalDiagnosisIsNotValid
-            );
-        });
-      }
-    }
-
-    if (referForm.controls['refrredToAdditionalServiceList'].value !== null) {
-      if (
-        referForm.controls['refrredToAdditionalServiceList'].value.length > 0
-      ) {
-        if (referForm.controls['referralReason'].errors) {
-          required.push(this.currentLanguageSet.Referdetails.referralReason);
-        }
-      } else if (referForm.controls['referredToInstituteName'].value !== null) {
-        if (referForm.controls['referralReason'].errors) {
-          required.push(this.currentLanguageSet.Referdetails.referralReason);
-        }
-      }
-    } else if (referForm.controls['referredToInstituteName'].value !== null) {
-      if (this.visitCategory === 'FP & Contraceptive Services') {
-        if (referForm.controls['referralReasonList'].errors) {
-          required.push(this.currentLanguageSet.Referdetails.referralReason);
-        }
-      } else {
-        if (referForm.controls['referralReason'].errors) {
-          required.push(this.currentLanguageSet.Referdetails.referralReason);
-        }
-      }
-    }
-
-    // For quick consult doctor flow, ensure at least one prescription exists
-    if (this.attendantType === 'doctor') {
-      try {
-        const quickConsultCaseRecordForm = <FormGroup>(
-          this.patientMedicalForm.controls['patientCaseRecordForm']
-        );
-        const prescription =
-          quickConsultCaseRecordForm && quickConsultCaseRecordForm.controls
-            ? quickConsultCaseRecordForm.controls['drugPrescriptionForm']
-            : null;
-        if (prescription) {
-          let prescribedDrugs =
-            prescription.value && prescription.value.prescribedDrugs
-              ? prescription.value.prescribedDrugs
-              : [];
-          prescribedDrugs = prescribedDrugs.filter((d: any) => !!d.createdBy);
-          if (!prescribedDrugs || prescribedDrugs.length === 0) {
-            required.push(
-              this.currentLanguageSet?.Prescription?.prescriptionRequired ||
-                'Please add at least one prescription'
-            );
-          }
-        }
-      } catch (err) {
-        console.warn(
-          'Error validating quick consult prescription presence',
-          err
-        );
-      }
-    }
-
-    if (required.length) {
-      this.confirmationService.notify(
-        this.currentLanguageSet.alerts.info.mandatoryFields,
-        required
-      );
-      this.resetSpinnerandEnableTheSubmitButton();
-      return 0;
-    } else {
-      return 1;
-    }
   }
 
   /**
@@ -3355,170 +2318,14 @@ export class WorkareaComponent
     }
   }
   checkNCDScreeningHistory(historyForm: any) {
-    const required = [];
-
-    let count = 0;
-    const familyDiseaseList =
-      historyForm.controls.patientHistoryForm.controls.familyHistory.controls
-        .familyDiseaseList.value;
-    familyDiseaseList.forEach((element: any) => {
-      if (
-        element.diseaseType !== null &&
-        element.deleted === false &&
-        element.diseaseType.diseaseType === 'Diabetes Mellitus'
-      ) {
-        count++;
-      }
-    });
-    if (this.beneficiaryAge < 30) {
-      count++;
-    }
-
-    if (count === 0) {
-      required.push(
-        this.currentLanguageSet.pleaseSelectDiabetesMellitusInFamilyHistory
-      );
-    }
-    let familyMember = 0;
-    const familyDiseasesList =
-      historyForm.controls.patientHistoryForm.controls.familyHistory.controls
-        .familyDiseaseList.value;
-    let familyDiseasesLength = familyDiseasesList.length;
-    for (let element = 0; element < familyDiseasesList.length; element++) {
-      if (
-        familyDiseasesList[element].diseaseType !== null &&
-        familyDiseasesList[element].deleted === false
-      ) {
-        if (
-          familyDiseasesList[element].familyMembers !== null &&
-          familyDiseasesList[element].familyMembers.length > 0
-        ) {
-          familyMember++;
-        }
-      } else {
-        familyDiseasesLength--;
-      }
-    }
-    if (familyMember !== familyDiseasesLength) {
-      required.push(this.currentLanguageSet.familyMemberInFamilyHistory);
-    }
-
-    if (required.length) {
-      this.confirmationService.notify(
-        this.currentLanguageSet.alerts.info.mandatoryFields,
-        required
-      );
-      return 0;
-    } else {
-      return 1;
-    }
+    return this.workareaValidation.checkNCDScreeningHistory(historyForm, this);
   }
 
   checkPastObstericHistory(generalOPDHistory: any) {
-    const vitalsForm = <FormGroup>(
-      generalOPDHistory.controls['patientHistoryForm']
+    return this.workareaValidation.checkPastObstericHistory(
+      generalOPDHistory,
+      this
     );
-    const pregForm1 = <FormGroup>vitalsForm.controls['pastObstericHistory'];
-    const pregForm2 = <FormGroup>pregForm1.controls['pastObstericHistoryList'];
-    const historyForm = <FormGroup>(
-      generalOPDHistory.controls['patientHistoryForm']
-    );
-    const required = [];
-    if (pregForm2.controls) {
-      const score1 = Number(pregForm2.controls['length']);
-      for (let i = 0; i < score1; i++) {
-        const pregForm3 = <FormGroup>pregForm2.controls[i];
-        if (
-          pregForm3.controls['pregOutcome'].value &&
-          pregForm3.controls['pregOutcome'].value.pregOutcome === 'Abortion'
-        ) {
-          if (
-            pregForm3.controls['abortionType'].value &&
-            pregForm3.controls['abortionType'].value.complicationValue ===
-              'Induced' &&
-            pregForm3.controls['typeofFacility'].errors
-          ) {
-            required.push(
-              this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                .typeofFacility +
-                '-' +
-                this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                  .orderofPregnancy +
-                ' ' +
-                pregForm3.value.pregOrder
-            );
-          }
-          if (pregForm3.controls['postAbortionComplication'].errors) {
-            required.push(
-              this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                .complicationPostAbortion +
-                '-' +
-                this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                  .orderofPregnancy +
-                ' ' +
-                pregForm3.value.pregOrder
-            );
-          }
-          if (pregForm3.controls['abortionType'].errors) {
-            required.push(
-              this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                .typeOfAbortion +
-                '-' +
-                this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                  .orderofPregnancy +
-                ' ' +
-                pregForm3.value.pregOrder
-            );
-          }
-          if (pregForm3.controls['pregDuration'].errors) {
-            required.push(
-              this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                .noOfcompletedWeeks +
-                '-' +
-                this.currentLanguageSet.historyData.opdNCDPNCHistory.obstetric
-                  .orderofPregnancy +
-                ' ' +
-                pregForm3.value.pregOrder
-            );
-          }
-        }
-      }
-    }
-    const personalHistory = historyForm.controls['personalHistory'];
-    const allergyList = personalHistory.value.allergicList;
-
-    let snomedTermNotMapped = false;
-
-    if (allergyList.length > 0) {
-      for (let i = 0; i < allergyList.length; i++) {
-        if (allergyList[i].allergyType !== null) {
-          if (
-            allergyList[i].snomedCode === null &&
-            allergyList[i].snomedTerm !== null
-          ) {
-            snomedTermNotMapped = true;
-          } else if (
-            allergyList[i].snomedCode !== null &&
-            allergyList[i].snomedTerm === null
-          ) {
-            snomedTermNotMapped = true;
-          }
-        }
-      }
-    }
-
-    if (snomedTermNotMapped) {
-      required.push(this.currentLanguageSet.allergyNameIsNotValid);
-    }
-    if (required.length) {
-      this.confirmationService.notify(
-        this.currentLanguageSet.alerts.info.mandatoryFields,
-        required
-      );
-      return 0;
-    } else {
-      return 1;
-    }
   }
 
   updatePatientExamination() {
